Update Remix to use Vite
