+++ app/components/ItemSelect.tsx
@@ -1,4 +1,4 @@
-import type { SlimItem } from "~/utils";
+import type { SlimItem } from "~/utils.server";
 import { itemToString } from "~/utils";
 
 import ItemName from "./ItemName";
