+++ vite.config.ts
@@ -1,0 +1,12 @@
+import { vitePlugin as remix } from "@remix-run/dev";
+import { installGlobals } from "@remix-run/node";
+import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from "vite";
+
+installGlobals();
+
+export default defineConfig({
+  plugins: [remix({
+    ignoredRouteFiles: ["**/.*"]
+  }), tsconfigPaths()],
+});
